zipkin-ui: fix jumpiness on "more info" button (#1847)
The "more info" button on the span detail view is all jumpy. Upon clicking the button, the appearing table gets squished together.

This happens because the `display` CSS property gets set to `block`. Where in this case it should really be `table`.

We can avoid the issue by wrapping the table in an extra div, as the div does not mind the `display: block`.